requirements.txt

python-telegram-bot==13.15
requests
pillow
pytesseract
speechrecognition
PyAudio
python-dotenv
ffmpeg-python
opencv-python-headless
numpy
PyMuPDF
python-docx
reportlab
python-bidi
arabic-reshaper
docx
fitz
opencv-python
telegram
pytesseract
python-dotenv
pytesseract
python-dotenv
